增加自定义错误, 增加数据库初始化操作, 增加jwt依赖注入

+++ services/admin.py
@@ -1,5 +1,6 @@
-from typing import Optional
+from typing import Optional, List
 from beanie import Document, Indexed
+from beanie import PydanticObjectId
 
 
 class Admin(Document):
@@ -11,20 +12,32 @@ class Admin(Document):
         collection = "admin"
 
 
-def get_admin_by_id(id: str):
-    return Admin.find_one(Admin.id == id)
+async def get_admin_by_id(id: PydanticObjectId):
+    admin = await Admin.find_one(Admin.id == id)
+    return admin
 
-def get_admin_by_email(email: str):
-    return Admin.find_one(Admin.email == email)
+async def get_admin_by_email(email: str):
+    admin = await Admin.find_one(Admin.email == email)
+    return admin
 
-def get_admins(skip: int = 0, limit: int = 100, sort_by: str = "email", order: str = "asc"):
-    return Admin.find().skip(skip).limit(limit).sort(sort_by, order)
+async def get_admins(skip: int = 0, limit: int = 100, sort_by: str = "email", order: str = "asc") -> List[Admin]:
+    admins = await Admin.find().skip(skip).limit(limit).sort(sort_by, order).to_list()
+    return admins
 
-def create_admin(admin: Admin):
-    return admin.save()
+async def create_admin(admin: Admin) -> Admin:
+    admin_ = await admin.create()
+    return admin_
 
-def update_admin(admin: Admin):
-    return admin.save()
+async def update_admin(id: PydanticObjectId, admin: Admin) -> Admin:
+    admin_ = await Admin.find_one(Admin.id == id)
+    values = admin.dict(exclude_unset=True)
+    for key, value in values.items():
+        setattr(admin_, key, value)
+    await admin_.save()
+    return admin_
 
-def delete_admin(admin: Admin):
-    return admin.delete()
+async def delete_admin(id: PydanticObjectId) -> bool:
+    admin = await Admin.find_one(Admin.id == id)
+    if admin:
+        await admin.delete()
+        return True
